batch: Extract source buffer loading
The batch package now has a source-cache module, while source snapshot
construction still owns the session-start buffer lookup used to populate
source commits.

The project does not provide a named boundary for reading file content from
abort snapshots, stash commits, and baseline commits. That makes
source_snapshots responsible for both locating input bytes and publishing
source commits.

This commit addresses that by moving the session-start buffer readers into
batch.source_buffers and by having source snapshot construction consume
them through private imports.

Subsequent test commits will pin direct buffer imports and guard against
old-location exports.
